requirements.txt
streamlit==1.33.0
torch==2.2.2
torchvision==0.17.2
timm==0.9.12
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==10.2.0
matplotlib==3.8.3
pywavelets==1.5.0